fix for names with keyword prefix affecting indent
For example:
if condition == true
  endGame()
end

would not indent then endOfGame() line correctly.

This still doesn't fix indentkeys indenting this wrong, but the same bug
affects Ruby, so it a systematic issue.

Thanks to:
Julien Cornebiese for bringing this to my attention.
and https://code.google.com/p/vim/issues/detail?id=280 for the general issue.